+++ CHANGELOG.md
@@ -7,6 +7,18 @@ and this project follows versions of format `{year}.{month}.{patch_number}`.
 
 ## [Unreleased]
 
+### Changed
+
+- [BREAKING] Changed the output of `read_results` to return a stronger-typed `CalibrationResults` 
+  pydantic model instance
+- [BREAKING] Renamed `read_session_results` to `read_results`
+- [BREAKING] Renamed `extract_bcc_params` to `generate_calib_seed_file`
+- [BREAKING] Changed `extract_bcc_params` (i.e. `generate_calib_seed_file`) to return no value
+  and only work with writing to a toml file. If one wants the results in dict or json form,
+  they can use the output of `read_results` and call `model_dumps` and `model_dumps_json` on it.
+- Added a `session_only` parameter on `read_results` to allow getting all results in redis regardless
+  of session if `session_only=True`
+
 ### Fixed
 
 - Removed the hard-coded fixes for CZ local phases for reverse phase qubits